GPUresamp module updates (#325)
* GPUresamp module updates

  * modify the gpu resamp code to produce results consistent with cpu
  * add the gpuresamp module to runFineResamp.py in TopsProc, which can be loaded topsApp.py
  * also fine-tune the sinc interpolation in cpu code: the sinc coef center value and normalization

+++ components/zerodop/GPUresampslc/include/GPUresamp.h
@@ -6,6 +6,6 @@
 #ifndef GPU_RESAMP_H
 #define GPU_RESAMP_H
 
-void runGPUResamp(double*,int*,void*,void*,double*,double*,double*,double*,double*,double*,double*,float*);
+void runGPUResamp(double*,int*,void*,void*,float*,float*,double*,double*,double*,double*,double*,float*);
 
 #endif
